reorganize equation logic
