+++ src/equations/constants.py
@@ -1,0 +1,11 @@
+"""
+Constants needed for calculating the electric field.
+"""
+
+import numpy as np
+
+EPSILON_NOUGHT = 1 / (4 * np.pi * 10**-7 * 299792458**2)
+COULOMB_CONSTANT = 1 / (4 * np.pi * EPSILON_NOUGHT)
+"""
+Coulomb's constant, in N m^2 / C^2.
+"""
